ASI-289 rename test_record to marc_record; delete duplicate class; create test task that accepts bib_id as argument to run MARC processing
